Deploy feature branches to staging or feature envs
When working on a non-trivial feature, it is often desirable to be able
to push the unmerged feature branch to the staging environment (or a
feature-specific environment) without merging it to master (on many
projects master is considered "ready to deploy").

This change alters the deploy subcommand for non-production
environments to push a non-master branch to the desired Heroku
environment's master branch (git push staging will actually push your
feature branch to a useless feature branch that doesn't get deployed on
staging) so that feature branches can be evaluated
without requiring a merge to master that might then have to be undone or
accidentally deployed to production.